Fixed espada MD, centrado indicador minimapa, removido ActualizarMini… (#463)
* Fixed espada MD, centrado indicador minimapa, removido ActualizarMiniMapa del main loop y reducida velocidad de danio sobre el char

* Codigo faltante...

+++ CODIGO/Red/Protocol.bas
@@ -11391,7 +11391,7 @@ Private Sub HandleCreateDamage()
         ' Leemos el ID del paquete.
         .ReadByte
      
-        Call mDx8_Dibujado.Damage_Create(.ReadByte(), .ReadByte(), 0, .ReadInteger(), .ReadByte())
+        Call mDx8_Dibujado.Damage_Create(.ReadByte(), .ReadByte(), 0, .ReadLong(), .ReadByte())
      
     End With
